requirements.txt
aiogram
psycopg2-binary
colorama
